+++ src/main/java/fr/tropweb/miningmanager/commands/Mining.java
@@ -3,7 +3,6 @@ package fr.tropweb.miningmanager.commands;
 import fr.tropweb.miningmanager.Utils;
 import fr.tropweb.miningmanager.commands.struct.CommandManager;
 import fr.tropweb.miningmanager.commands.struct.SubCommand;
-import fr.tropweb.miningmanager.data.Settings;
 import fr.tropweb.miningmanager.engine.Engine;
 import fr.tropweb.miningmanager.pojo.MiningTask;
 import fr.tropweb.miningmanager.pojo.PlayerLite;
@@ -11,7 +10,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.Chunk;
 import org.bukkit.command.CommandException;
 import org.bukkit.entity.Player;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitTask;
 
 import java.util.Arrays;
@@ -21,26 +19,24 @@ import static fr.tropweb.miningmanager.commands.struct.CommandManager.*;
 
 public class Mining implements SubCommand {
     private final Engine engine;
-    private final Plugin plugin;
 
     public Mining(Engine engine) {
         this.engine = engine;
-        this.plugin = this.engine.getPlugin();
     }
 
     @Override
     public void onCommand(final Player player, final CommandManager attribute) {
 
-        // get player data
+        // get player data from memory
         final PlayerLite playerLite = this.engine.getPlayerEngine().getPlayerLite(player);
 
-        // get mining task
+        // get mining task from player
         final MiningTask miningTask = playerLite.getMiningTask();
 
-        // get chunk from player
+        // get chunk of player
         final Chunk chunk = player.getLocation().getChunk();
 
-        // check if player want to stop
+        // check if player want to stop the current mining
         if (attribute == STOP_MINING) {
 
             // check if mining task running
@@ -51,14 +47,11 @@ public class Mining implements SubCommand {
             // stop mining task
             miningTask.stopMiningTask();
 
-            // inform player
+            // inform player if the task end properly
             Utils.green(player, "Mining has been stopped.");
-
-            // end
-            return;
         }
 
-        // return to the player the number of precious resource left
+        // check if the player want the number of precious resource left
         else if (attribute == SHOW_MINING) {
 
             // check if mining task running
@@ -66,50 +59,75 @@ public class Mining implements SubCommand {
                 throw new CommandException("Mining is not running.");
             }
 
-            // if there is block to mine
+            // check if player have mining block left
             if (miningTask.hasBlockToMine()) {
 
-                // scan for the player
-                this.engine.getChunkEngine().onCommandInChunkOfPlayer(player, miningTask.getChunk());
+                // start free scan for the player
+                Scan.scanChunkOfPlayer(this.engine, player, miningTask.getChunk());
+            }
+        }
+
+        // enter to the mining case
+        else {
+
+            // check if mining task running
+            if (miningTask.hasMiningTask()) {
+
+                // player cannot have two mining task
+                throw new CommandException("You cannot run two mining at the same time.");
             }
 
-            // end
-            return;
+            // check if other player already have start mining task in this chunk
+            else if (this.engine.getPlayerEngine().isChunkAlreadyMined(chunk)) {
+
+                // two mining task cannot start at the same chunk
+                throw new CommandException("Another player is already mining out this chunk.");
+            }
+
+            // clear block list of the player
+            miningTask.getBlockToMine().clear();
+
+            // list the precious block of the chunk and add it to the player block list
+            miningTask.getBlockToMine().addAll(this.engine.getChunkEngine().getBlockFromChunk(chunk));
+
+            // check if there is block to mine
+            if (!miningTask.hasBlockToMine()) {
+
+                // we should not have empty block list
+                throw new CommandException("There is no precious block to mine.");
+            }
+
+            // check if economy plugin is enabled
+            if (this.economyEnabled()) {
+
+                // check if player can pay and take money
+                if (this.engine.getEconomyPlugin().takeMoney(player, this.engine.getSettings().getMiningPrice())) {
+
+                    // inform player about the money taken
+                    Utils.green(player, "You have spend %s$ to start the mining.", this.engine.getSettings().getMiningPrice());
+                } else {
+
+                    // if player don't have enough money
+                    throw new CommandException("You don't have enough money to start mining");
+                }
+            }
+
+            // The player have 15 (by default) to select a chest
+            final BukkitTask task = Bukkit.getScheduler().runTaskLater(
+                    this.engine.getPlugin(),
+                    () -> this.stopMining(player, miningTask),
+                    this.engine.getSettings().getTickMiningTimeout()
+            );
+
+            // the task is keep in memory
+            miningTask.setMiningTask(task);
+
+            // keep in mind the chunk
+            miningTask.setChunk(chunk);
+
+            // inform the player about the delay
+            Utils.green(player, "Right click to the chest to start the mining. This action will be close in %ss.", this.engine.getSettings().getMiningTimeout());
         }
-
-        // if player already have mining task
-        else if (miningTask.hasMiningTask()) {
-            throw new CommandException("You cannot run two mining at the same time.");
-        }
-
-        // if other player already have start mining task
-        else if (this.engine.getPlayerEngine().isChunkAlreadyMined(chunk)) {
-            throw new CommandException("Another player is already mining out this chunk.");
-        }
-
-        // run scan of the chunk and collect info
-        this.engine.getChunkEngine().onCommandInChunkOfPlayer(player, chunk, miningTask.getBlockToMine());
-
-        // if there is block
-        if (!miningTask.hasBlockToMine()) {
-            Utils.red(player, "There is no precious block to mine.");
-            return;
-        }
-
-        // get settings
-        final Settings settings = this.engine.getSettings();
-
-        // task to leave in 15s
-        final BukkitTask task = Bukkit.getScheduler().runTaskLater(this.engine.getPlugin(), () -> this.stopMining(player, miningTask), settings.getTickMiningTimeout());
-
-        // save task
-        miningTask.setMiningTask(task);
-
-        // keep in mind the chunk
-        miningTask.setChunk(chunk);
-
-        // inform the player
-        Utils.green(player, "Right click to the chest to start the mining. This action will be close in %ss.", settings.getMiningTimeout());
     }
 
     @Override
@@ -128,14 +146,26 @@ public class Mining implements SubCommand {
     }
 
     private void stopMining(final Player player, final MiningTask miningTask) {
-        // check if mining task
-        if (miningTask.hasMiningTask()) {
 
-            // stop current task
-            miningTask.stopMiningTask();
+        // check if economy plugin is enabled
+        if (this.economyEnabled()) {
 
-            // inform
-            Utils.red(player, "You mining task has been abort.");
+            // money get back to the player
+            this.engine.getEconomyPlugin().giveMoney(player, this.engine.getSettings().getMiningPrice());
+
+            // inform about money
+            Utils.green(player, "You retrieve %s$.", this.engine.getSettings().getMiningPrice());
         }
+
+        // stop current task
+        miningTask.stopMiningTask();
+
+        // inform about abort
+        Utils.red(player, "You mining task has been abort.");
+    }
+
+    private boolean economyEnabled() {
+        // check if economy plugin is enabled and if there is price
+        return this.engine.getEconomyPlugin().isEnabled(this.engine.getSettings().getMiningPrice());
     }
 }