+++ src/main/java/fr/tropweb/miningmanager/listeners/PlayerEventHandler.java
@@ -1,7 +1,7 @@
 package fr.tropweb.miningmanager.listeners;
 
 import fr.tropweb.miningmanager.Utils;
-import fr.tropweb.miningmanager.data.Settings;
+import fr.tropweb.miningmanager.commands.Scan;
 import fr.tropweb.miningmanager.engine.Engine;
 import fr.tropweb.miningmanager.pojo.MiningTask;
 import fr.tropweb.miningmanager.pojo.PlayerLite;
@@ -16,70 +16,95 @@ import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.scheduler.BukkitTask;
 
 public class PlayerEventHandler implements Listener {
-
     private final Engine engine;
 
-    public PlayerEventHandler(Engine engine) {
+    public PlayerEventHandler(final Engine engine) {
         this.engine = engine;
     }
 
     @EventHandler(ignoreCancelled = true, priority = EventPriority.LOWEST)
     public void onPlayerMove(final PlayerMoveEvent event) {
+
+        // event is expected
         if (event == null) return;
+
+        // player is expected
         if (event.getPlayer() == null) return;
 
-        final PlayerLite playerLite = this.engine.getPlayerEngine().getPlayerLite(event.getPlayer());
+        // constant list
+        final Player player = event.getPlayer();
+        final PlayerLite playerLite = this.engine.getPlayerEngine().getPlayerLite(player);
 
-        // if auto scan, if chunk different
-        if (playerLite.isAutoScan() && !event.getFrom().getChunk().equals(event.getTo().getChunk()))
-            this.engine.getChunkEngine().onCommandInChunkOfPlayer(event.getPlayer(), event.getTo().getChunk());
+        // check if auto scan = true, and if chunk is different
+        if (playerLite.isAutoScan() && !event.getFrom().getChunk().equals(event.getTo().getChunk())) {
+
+            // check if economy plugin is enabled
+            if (this.engine.getEconomyPlugin().isEnabled(this.engine.getSettings().getScanPrice())) {
+
+                // check if the money of the player has been taken
+                if (!this.engine.getEconomyPlugin().takeMoney(player, this.engine.getSettings().getScanPrice())) {
+
+                    // unset auto scan
+                    playerLite.setAutoScan(false);
+
+                    // inform player that he don't have enough money
+                    Utils.red(player, "You don't have enough money to continue the auto scan.");
+
+                    // stop process
+                    return;
+                }
+            }
+
+            // scan the chunk's player
+            Scan.scanChunkOfPlayer(this.engine, player, event.getTo().getChunk());
+        }
     }
 
     @EventHandler(ignoreCancelled = true, priority = EventPriority.LOWEST)
     public void onPlayerSelectChest(final PlayerInteractEvent event) {
+
+        // event is expected
         if (event == null) return;
 
+        // player is expected
         final Player player = event.getPlayer();
         if (player == null) return;
 
+        // clicked block is expected
         final Block container = event.getClickedBlock();
         if (container == null) return;
 
-        // get player from memory
+        // constant list
         final PlayerLite playerLite = this.engine.getPlayerEngine().getPlayerLite(player);
-
-        // get mining task from player
         final MiningTask miningTask = playerLite.getMiningTask();
 
-        // if have the end task and if he doesn't choose chest
+        // check if there is mining task running without selected chest
         if (miningTask.hasMiningTask() && !miningTask.hasMiningChest()) {
 
-            // if it's chest
+            // check if the selected block is a chest
             if (this.engine.getBlockEngine().isContainer(container)) {
 
-                // add chest to player
+                // configure chest for the player
                 miningTask.setMiningChest(container);
 
-                // cancel task
+                // cancel the starting task
                 miningTask.getMiningTask().cancel();
 
-                // get settings
-                final Settings settings = this.engine.getSettings();
-
-                // create task
+                // create mining task
                 final BukkitTask task = Bukkit.getScheduler().runTaskTimer(
                         this.engine.getPlugin(),
                         () -> this.engine.getMiningEngine().startMining(player),
-                        settings.getTickMiningStart(),
-                        settings.getTickMiningInterval());
+                        this.engine.getSettings().getTickMiningStart(),
+                        this.engine.getSettings().getTickMiningInterval()
+                );
 
-                // save task
+                // save task to the player
                 miningTask.setMiningTask(task);
 
-                // inform player
-                Utils.green(player, "Mining starting... loot every %ss.", settings.getMiningInterval());
+                // inform player that the mining start has been started
+                Utils.green(player, "Mining starting... loot every %ss.", this.engine.getSettings().getMiningInterval());
 
-                // cancel the click
+                // cancel the click event to avoid to open the chest
                 event.setCancelled(true);
             }
         }